server/requirements.txt
fastapi>=0.115.0
huggingface_hub
openai
requests>=2.31.0,<3.0.0
pydantic>=2.7.0,<3.0.0
fastapi>=0.115.0,<1.0.0
uvicorn>=0.30.0,<1.0.0
openenv-core==0.2.3